requirements.txt
aiogram==2.14
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.28.0
attrs==23.1.0
Babel==2.13.1
certifi==2023.7.22
charset-normalizer==3.3.2
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
psutil==5.9.6
psycopg2==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
yarl==1.9.2
